fiz crossover, mutação, aparentimente está terminado

+++ src/main/java/br/uva/algoritmoGenetico/dominio/Avaliacao.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class Avaliacao {
 
     public static void avaliacao(List<Individuo> populacao, Equacao equacao){
-        int somatorioAvaliacao = 0;
+        double somatorioAvaliacao = 0;
         for(Individuo individuo : populacao){
             individuo.avaliaIndividuo(equacao);
             somatorioAvaliacao += individuo.getValorAvaliacao();
